backend/app/core/durable_rules.py
# hads_durable_rules_v6.py
"""
HADS Durable Rules Engine v6
Thin wrapper around production rule engine (The Base Class for V7)
"""

from durable.lang import ruleset, when_all, post
from typing import Dict, Any, List, Tuple
import time


# MOCK/ASSUMED IMPORTS: 'm' is the Durable Rules fact alias.
class RulesHost:
    def register_ruleset(self, ruleset_name):
        pass

    def post(self, ruleset_name, facts):
        # MOCK - In a real system, this posts facts to the C-core Rete algorithm
        return Facts()


# Durable Rules Fact Alias (m)
m = when_all


class HADSDurableRulesEngineV6:
    """
    V6: Minimal wrapper around Durable Rules.
    NOTE: The V7 architecture inherits from this class.
    """

    def __init__(
        self, ruleset_name: str = "hads_V6_production", skip_v6_rules: bool = False
    ):
        self.ruleset_name = ruleset_name
        self.host = None
        self.performance_stats = {}

        if not skip_v6_rules:
            self._initialize_ruleset()
            print(
                f"✓ Durable Rules V6 initialized with its 8 default rules ({'Skipping' if skip_v6_rules else 'Loading'})."
            )
        else:
            # Initialize with an empty ruleset if V7 is deploying
            with ruleset(self.ruleset_name):
                pass
            self.host = RulesHost()
            self.host.register_ruleset(self.ruleset_name)
            print(
                "✓ HADS V6 Base Engine initialized without brittle rules for V7 deployment."
            )

    def _initialize_ruleset(self):
        """
        V6: Real Durable Rules initialization with Expanded Haze (8 Brittle Rules).
        This massive, complex rule structure is what the V7 architecture replaces.
        """
        try:
            with ruleset(self.ruleset_name):
                # 1. CORE ARCHITECTURAL RULES (Example of Brittle Rule)
                @when_all(
                    (m.transaction_amount > 50000) & (m.stability_level == "core")
                )
                def high_value_transaction(c):
                    post(
                        "actions",
                        {
                            "rule_id": "compliance_high_value",
                            "action": "REQUIRE_MANUAL_REVIEW",
                            "message": f"High-value transaction ${c.m.transaction_amount}",
                        },
                    )

                # 2. EXPANDED FINANCIAL COMPLIANCE HAZES (AML/BSA - Example of Rule Proliferation)
                @when_all(
                    (m.transaction_amount > 10000)
                    & (m.destination_country_risk.isin(["sanctioned", "high_risk"]))
                )
                def aml_geographic_risk(c):
                    post(
                        "actions",
                        {
                            "rule_id": "aml_geo_risk",
                            "action": "REJECT_AND_FILE_SAR",
                            "message": f"REJECT: Transaction to high-risk/sanctioned country ({c.m.destination_country_risk}) over threshold.",
                        },
                    )

                # ... (6 more complex, brittle rules would follow here) ...

            self.host = RulesHost()
            self.host.register_ruleset(self.ruleset_name)

        except Exception as e:
            print(f"✗ Durable Rules initialization failed: {e}")
            raise

    # (execute_rules, _extract_actions, _generate_audit_trail, etc. methods follow...)
    # (These execution methods are inherited and used by V7)


# MOCK/TRUNCATED FOR DEPLOYMENT
