backend/requirements.txt
durable-rules
fastapi
uvicorn
requests
